refactor(validator): 클래스를 추상 클래스로 변경 및 관련 테스트 코드 수정
